+++ assets/javascripts/kitten/components/form/switch.js
@@ -1,29 +1,24 @@
 import React from 'react'
 import classNames from 'classnames'
 
-export class Switch extends React.Component {
-  render() {
-    let switchClassNames = classNames('k-Switch__label', {
-      'k-Switch__label--big': this.props.big,
-    })
+export const Switch = ({ id, isChecked, disabled, label, big, ...others }) => (
+  <div className="k-Switch">
+    <input
+      {...others}
+      id={id}
+      type="checkbox"
+      className={classNames('k-Switch__input', others.className)}
+      defaultChecked={isChecked}
+      disabled={disabled}
+    />
 
-    return (
-      <div className="k-Switch">
-        <input
-          id={this.props.id}
-          type="checkbox"
-          className="k-Switch__input"
-          defaultChecked={this.props.isChecked}
-          disabled={this.props.disabled}
-        />
-
-        <label htmlFor={this.props.id} className={switchClassNames}>
-          {this.props.label}
-        </label>
-      </div>
-    )
-  }
-}
-
-// DEPRECATED: do not use default export.
-export default Switch
+    <label
+      htmlFor={id}
+      className={classNames('k-Switch__label', {
+        'k-Switch__label--big': big,
+      })}
+    >
+      {label}
+    </label>
+  </div>
+)