+++ src/components/form/field/index.js
@@ -29,6 +29,8 @@ var _select = require("./components/select");
 
 var _checkbox = require("./components/checkbox");
 
+var _password = require("./components/password");
+
 var Field =
 /*#__PURE__*/
 function (_Component) {
@@ -64,3 +66,4 @@ Field.RadioButtonSet = _radioButtonSet.FieldRadioButtonSet;
 Field.Select = _select.FieldSelect;
 Field.Checkbox = _checkbox.FieldCheckbox;
 Field.ErrorMessage = _error.FieldError;
+Field.Password = _password.FieldPassword;